Refactored to format I could understand better
