+++ static/swiss.css
@@ -1,0 +1,31 @@
+body {
+    font-family: Arial, sans-serif;
+}
+
+.tab-content {
+    display: none;
+    padding: 20px;
+}
+
+table {
+    width: 100%;
+    border-collapse: collapse;
+}
+
+table, th, td {
+    border: 1px solid #ddd;
+    text-align: left;
+    padding: 8px;
+}
+
+th {
+    background-color: #f2f2f2;
+}
+
+form {
+    margin-bottom: 20px;
+}
+
+button {
+    cursor: pointer;
+}
